pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-intelligence"
version = "0.1.0"
description = "CS:GO Analytics AI — RAG pipeline with LlamaIndex, FastAPI, and pgvectorscale."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "kaggle>=1.7.4.5",
  "minio>=7.2.18",
  "python-dotenv>=1.1.0",
  "fastapi>=0.115",
  "uvicorn[standard]>=0.34",
  "llama-index-core>=0.14",
  "llama-index-vector-stores-postgres>=0.7",
  "llama-index-llms-openai>=0.6",
  "llama-index-llms-anthropic>=0.10",
  "llama-index-llms-ollama>=0.9",
  "llama-index-embeddings-openai>=0.5",
  "llama-index-embeddings-ollama>=0.8",
  "llama-index-embeddings-voyageai>=0.5",
  "asyncpg>=0.31",
  "psycopg2-binary>=2.9",
  "sqlalchemy[asyncio]>=2.0",
  "pgvector>=0.4",
  "pydantic>=2.0",
  "pydantic-settings>=2.0",
  "structlog>=24.4",
  "opentelemetry-api>=1.28",
  "opentelemetry-sdk>=1.28",
  "opentelemetry-exporter-otlp-proto-http>=1.28",
  "opentelemetry-instrumentation-fastapi>=0.49b0",
  "llama-index-observability-otel>=0.3",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5",
  "pytest-asyncio>=0.24",
  "httpx>=0.27",
]
ml = [
  "scikit-learn>=1.5",
  "pandas>=2.2",
  "numpy>=1.26",
]
mlops = [
  "mlflow>=2.18,<3",
]

[project.scripts]
bronze-import = "rag_intelligence.cli:main"
silver-transform = "rag_intelligence.silver_cli:main"
gold-transform = "rag_intelligence.gold_cli:main"
document-build = "rag_intelligence.document_cli:main"
embedding-ingest = "rag_intelligence.embedding_cli:main"
semantic-search = "rag_intelligence.search_cli:main"
run-audit = "rag_intelligence.run_audit_cli:main"
train-logreg = "rag_intelligence.train_cli:main_logreg"
train-histgbt = "rag_intelligence.train_cli:main_histgbt"
train-baseline = "rag_intelligence.train_cli:main_baseline"
embed-docs = "rag_intelligence.embed_docs_cli:main"
rag-api = "rag_intelligence.api.main:run"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py313"
line-length = 99
src = ["src", "tests"]

[tool.ruff.lint]
select = [
  "E",    # pycodestyle errors
  "F",    # pyflakes
  "I",    # isort
  "UP",   # pyupgrade
  "B",    # bugbear
  "SIM",  # simplify
  "RUF",  # ruff-specific
]

[tool.pyright]
pythonVersion = "3.13"
venvPath = "."
venv = ".venv"
typeCheckingMode = "basic"

[tool.pytest.ini_options]
testpaths = ["tests"]
